Move files to src directory to prepare for TS
