+++ src/triggerable-branch.js
