+++ test/helpers/config-mock.js
@@ -1,6 +1,6 @@
 const fs = require('fs')
 const nock = require('nock')
-const { encodeContent } = require('../../lib/base64')
+const { encodeContent } = require('../../src/base64')
 
 function configFixture(fileName = 'config.yml') {
   return {
